Casted away of signed against unsigned warning.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@11145 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/import/mdl2spr/m2s_md2.cpp
@@ -380,7 +380,7 @@ bool Md2::WriteSPR(const char* spritename, float scaleMdl, int delayMdl,
         for (j = strlen(frames[i].name)-1; j > 1; j--)
           if (!isdigit(frames[i].name[j]))
             break;
-        if(strlen(frames[i].name) > j + 3)
+        if(strlen(frames[i].name) > (unsigned int) j + 3)
         {
             base_action=j + 2;
         }
